Fix command queue issue
