+++ schema/changelog-5.5.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog
+  xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+                      http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.4.xsd"
+  logicalFilePath="changelog-5.5">
+
+  <changeSet author="author" id="changelog-5.5">
+
+    <dropColumn tableName="tc_commands_queue" columnName="description" />
+
+  </changeSet>
+
+</databaseChangeLog>
