+++ schema/changelog-master.xml
@@ -35,5 +35,6 @@
   <include file="changelog-5.2.xml" relativeToChangelogFile="true" />
   <include file="changelog-5.3.xml" relativeToChangelogFile="true" />
   <include file="changelog-5.4.xml" relativeToChangelogFile="true" />
+  <include file="changelog-5.5.xml" relativeToChangelogFile="true" />
 
 </databaseChangeLog>
