+++ src/main/java/org/traccar/model/QueuedCommand.java
@@ -15,18 +15,19 @@
  */
 package org.traccar.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import org.traccar.storage.StorageName;
 
 import java.util.HashMap;
 
 @StorageName("tc_commands_queue")
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class QueuedCommand extends BaseCommand {
 
     public static QueuedCommand fromCommand(Command command) {
         QueuedCommand queuedCommand = new QueuedCommand();
         queuedCommand.setDeviceId(command.getDeviceId());
         queuedCommand.setType(command.getType());
-        queuedCommand.setDescription(command.getDescription());
         queuedCommand.setTextChannel(command.getTextChannel());
         queuedCommand.setAttributes(new HashMap<>(command.getAttributes()));
         return queuedCommand;
@@ -36,7 +37,7 @@ public class QueuedCommand extends BaseCommand {
         Command command = new Command();
         command.setDeviceId(getDeviceId());
         command.setType(getType());
-        command.setDescription(getDescription());
+        command.setDescription("");
         command.setTextChannel(getTextChannel());
         command.setAttributes(new HashMap<>(getAttributes()));
         return command;
